+++ cse_hub/users/forms.py
@@ -1,5 +1,8 @@
 from django.forms import ModelForm
 from .models import Profile
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
+from django import forms
 
 class ProfileUpdateForm(ModelForm):
 	class Meta:
@@ -8,4 +11,33 @@ class ProfileUpdateForm(ModelForm):
 		# widgets = {
 		# 	'password': forms.PasswordInput(),
 		# }
-		fields = []
+		fields = [
+			'profile_pic'
+		]
+
+class RegistrationForm(UserCreationForm):
+    email = forms.EmailField(required = True)
+    profile_pic = forms.ImageField(required=False)
+
+    class Meta:
+        model = User
+        fields = (
+            'username',
+            'first_name',
+            'last_name',
+            'email',
+            'password1',
+            'password2',
+            'profile_pic'
+        )
+
+    def save(self, commit=True):
+        user = super(RegistrationForm, self).save(commit=False)
+        user.first_name = self.cleaned_data['first_name']
+        user.last_name = self.cleaned_data['last_name']
+        user.email = self.cleaned_data['email']
+
+        if commit:
+            user.save()
+
+        return user
